Separate dseth tests between mint and redeem
